New: ~8% faster llama inference.
All the linear layers in the Llama model have `bias=False`, so let's not bother copying a bunch of zeroes to the output tensor in each layer when we can just zero-fill.

Testing time to generate 2047 tokens with 7B llama, --wbits 4 --groupsize 128, on an RTX 4090, not counting the time to load the model, generation time goes from 1m7.8s before to 1m2.4s after (92%).