Make sure to save the quantized model before bench/eval
Depending on the configuration of the quantization parameters and GPU hardware, it can take several hours to fully quantize the model. 

Unfortunately, benchmarking and model evaluation may increase vram usage and cause oom.
 
To prevent losing all the hard work, gone in a snap, followed by palm-to-face and coffee-in-our-lap moments, we must prioritize saving the quantized models over everything else.